This is a quick hack to get a CTCSS tone generator going on a sparkfun pro micro with an OakOLED display.
It uses a timer to generate the CTCSS tones but it can't use interrupts
for the button presses due to a lack of interrupts on all of the input
pins.  I may work around this with some diodes at some point to trigger
the interrupt to check things but .. well, slackness.

It outputs a square wave which is difficult to turn into a sine wave
over the potential PL tones range (60->260Hz.)

But hey! I haven't done this stuff in a few years.  It's fun!