Blank the screen after the eeprom write is done.
This OLED module generates a lot of RF noise that is leaking back into
the power / audio wires on my prototype.  Now, ideally I'd eliminate
this on the board by suitable application of inductors and filter
caps but I'm not yet there.

Blanking the screen causes the noise to almost disappear.
There's still a tiny amount, but that sounds like the CPU generated
noise as it's in a tight loop.